requirements.txt
streamlit==1.28.1
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
gTTS==2.3.0
accelerate==0.24.1
sentencepiece==0.1.99
protobuf==3.20.3